requirements.txt
adal==1.2.2
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.0
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
cycler==0.11.0
DateRangeParser==1.3.2
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Django==2.1.11
django-auth-adfs==1.2.0
django-cors-headers==3.1.1
django-cors-middleware==1.4.0
django-oauth-toolkit==1.2.0
django-rest-swagger==2.2.0
djangorestframework==3.10.3
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.3
fonttools==4.39.2
fqdn==1.5.1
fuzzywuzzy==0.18.0
idna==3.4
importlib-metadata==6.1.0
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.5
isoduration==20.11.0
itypes==1.2.0
jdcal==1.4.1
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.6
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.5
more-itertools==9.1.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.23.1
oauthlib==3.2.2
openapi-codec==1.3.2
openpyxl==3.0.3
packaging==23.0
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.20.0
python-json-logger==2.0.7
pytz==2022.7.1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.1
QtPy==2.3.0
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.2.2
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.6
zipp==0.6.0
#reference
#https://djangocentral.com/django-ajax-with-jquery/